Fixed buttons' look and language and added some further validation checks for datepickers.
When picking an end date first, for example, and then picking a start date, the start date cannot be after the end date.
I don't know why somebody might choose to do it in that order, but now, upper and lower bounds are restricted in each
date picker according to the date selected in the other one.